El plano de mesas: paginas, secciones y modo edicion
Maqueta aprobada por Jhon, .sql corrido, y aca esta la pantalla. Las mesas dejan
de ser una lista suelta con una etiqueta de salon pegada: ahora hay paginas
arriba (Salon, Mostrador, las que el cree) y adentro secciones en tarjetas de
color suave, con las mesas en cada una.

El modo edicion entra por el lapiz al final de la franja y se anuncia con una
cinta naranja: agregar y quitar mesas, moverlas de seccion, crear y renombrar
secciones y paginas, forma y tamano por mesa.

SIN EL .SQL EL PLANO SIGUE EXACTAMENTE COMO ANTES, agrupando por salon, y el
lapiz ni aparece. Publicar una pantalla que se queda en blanco esperando un .sql
es la falla silenciosa que este proyecto ya pago dos veces.

DONDE ME APARTE DE LA MAQUETA: decia "arrastrarla la mueve a otra seccion" y no
se hizo con arrastre. En el telefono el plano es un riel de 66 px que ademas
hace scroll, y arrastrar ahi es la receta de mover una mesa sin querer — y una
mesa movida por accidente manda el pedido al lugar equivocado. Se toca la mesa y
se elige la seccion en su ficha: un gesto que no se dispara solo.

Los tres candados son de los que protegen datos: una mesa con la cuenta abierta
no se quita ni se mueve, una seccion con mesas no se borra, una pagina con
secciones tampoco. Los tres avisan con el motivo. Y quitar una mesa no la borra:
la desactiva, porque tiene cuentas y ventas colgando.

Ademas se cambio el confirm() del navegador por `preguntar`, que es la ventana
de la casa: mezclar las dos formas de preguntar se nota al instante.

pruebas/lama-plano.mjs, 38 casos, con el .sql corrido y sin correr. Y uno de
ellos pasaba por la razon equivocada: el candado de la mesa con cuenta abierta
seguia en verde con el candado sacado, porque sin el aparece una confirmacion
que la prueba nunca contestaba, y no contestarla tambien impedia la escritura.
Ahora la prueba acepta lo que aparezca. Comprobado rompiendo el codigo a
proposito, en los dos sentidos.

Toda la bateria de Lama en verde: plano 38, mesas 36, mover 33, config 35,
ancho 26, cobrar 32, limpieza 48, precuenta 13, pc-riel 30.

+++ sql/2026-09-lama-plano-editable.sql
@@ -95,9 +95,16 @@ alter table public.mesas add  constraint mesas_tam_ok
 -- Cada sede estrena una página "Salón", y adentro una sección por cada salón
 -- que ya existía. Después las movés vos desde la app, que es para lo que
 -- sirve el modo edición.
+--
+-- ⚠️ CORREGIDO EL 2026-09-05: el `null` de `padre_id` va con `::bigint`.
+-- Uno supone que Postgres deduce el tipo de la columna de destino, y NO lo
+-- hace dentro de un `insert ... select`: ahí un `null` pelado nace como
+-- TEXTO, y la columna es bigint. Falla con "column padre_id is of type
+-- bigint but expression is of type text". Es la misma familia de error que
+-- §0.1.9: escribir de memoria en vez de decirlo explícito.
 -- ================================================================
 insert into public.lama_areas (sede, nombre, padre_id, orden, color)
-select distinct m.sede, 'Salón', null, 0, 'a'
+select distinct m.sede, 'Salón', null::bigint, 0, 'a'
   from public.mesas m
  where not exists (select 1 from public.lama_areas a
                     where a.sede = m.sede and a.padre_id is null);
